Upgrade VoxRuntime, regenerate IPC code, update AGENTS.md
- Bump vox to 901eeba (adds UnixAcceptor, removes vox_shm_ffi)
- Regenerate AppIPC.swift and ImeIPC.swift via bee-rpc codegen
- Simplify BeeIPCServer to use UnixAcceptor instead of manual accept loop
- Fix Int→UInt casts for cbindgen-generated FFI signatures
- Rewrite AGENTS.md for current project structure, add CLAUDE.md symlink

+++ swift/bee/Services/TranscriptionService.swift
@@ -127,9 +127,9 @@ final class TranscriptionService: @unchecked Sendable {
         var err: UnsafeMutablePointer<CChar>?
         let result: AsrFeedResult = samples.withUnsafeBufferPointer { buf in
             if finalizing {
-                return asr_session_feed_finalizing(session.ptr, buf.baseAddress, buf.count, &err)
+                return asr_session_feed_finalizing(session.ptr, buf.baseAddress, UInt(buf.count), &err)
             } else {
-                return asr_session_feed(session.ptr, buf.baseAddress, buf.count, &err)
+                return asr_session_feed(session.ptr, buf.baseAddress, UInt(buf.count), &err)
             }
         }
 
@@ -188,7 +188,7 @@ final class TranscriptionService: @unchecked Sendable {
 
         var err: UnsafeMutablePointer<CChar>?
         let result = samples.withUnsafeBufferPointer { buf in
-            asr_engine_transcribe_samples(engine, buf.baseAddress, buf.count, &err)
+            asr_engine_transcribe_samples(engine, buf.baseAddress, UInt(buf.count), &err)
         }
 
         if let err {
